refactor: use design system glass background
